Update build.yml permissions for pushing contents and packages
The build.yml file has been updated to modify the permissions for pushing contents and packages. Instead of setting the permissions to "write-all", the new permissions are set to "contents: write" and "packages: write". This change ensures that the necessary permissions are granted for pushing contents and packages in the workflow.